perbaikan dan menambahkan fitur loans dan navbar untuk layouts

+++ books/index_books.php
@@ -1,8 +1,8 @@
 <?php
 require_once __DIR__ . '/../connection.php';
 
-$query_books = "SELECT books.*, categories.category_name 
-FROM books 
+$query_books = "SELECT books.*, categories.category_name
+FROM books
 INNER JOIN categories ON books.category_id = categories.id";
 $query = mysqli_query($connect, $query_books);
 
@@ -24,13 +24,17 @@ $query = mysqli_query($connect, $query_books);
         }
 
         tbody,
-        tr,td {
+        tr,
+        td {
             padding: 4px;
         }
     </style>
 </head>
 
 <body>
+
+    <?php $base_path = '../' ?>
+    <?php include '../layouts/navbar.php'; ?>
     <h1>Halaman Buku</h1>
 
     <button><a href="create_books.php">Tambah Buku</a></button>
@@ -39,6 +43,7 @@ $query = mysqli_query($connect, $query_books);
     <table border="1">
         <thead>
             <th>No</th>
+            <th>Kode Buku</th>
             <th>Judul Buku</th>
             <th>Kategori</th>
             <th>Penulis</th>
@@ -56,6 +61,7 @@ $query = mysqli_query($connect, $query_books);
             ?>
                 <tr>
                     <td><?= $no++ ?></td>
+                    <td><?= $data['book_code'] ?></td>
                     <td><?= $data['title'] ?></td>
                     <td><?= $data['category_name'] ?></td>
                     <td><?= $data['author']  ?></td>
@@ -65,7 +71,7 @@ $query = mysqli_query($connect, $query_books);
                     <td style="width: 500px;"><?= $data['description'] ?></td>
                     <td>
                         <a href="edit_books.php?id=<?= $data['id'] ?>">Edit</a>
-                        <a href="delete_books.php?id=<?= $data['id'] ?>">Hapus</a>
+                        <a href="delete_books.php?id=<?= $data['id'] ?>" onclick="return confirm(' apakah yakin ingin menghapus buku ini?')">Hapus</a>
                     </td>
                 </tr>
             <?php endwhile; ?>
